backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.20
anthropic>=0.40.0
google-genai>=1.0.0
PyMuPDF==1.25.1
python-docx==1.1.2
aiofiles==24.1.0
pydantic[email]>=2.0.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
httpx==0.27.2
